backend/requirements.txt
# RakshakAI Backend Requirements
# Production-grade dependencies for scam call defense system

# ==========================================
# CORE FRAMEWORK
# ==========================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
uvloop==0.19.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ==========================================
# WEBSOCKET & REAL-TIME
# ==========================================
websockets==12.0
starlette==0.27.0
aiofiles==23.2.1
aioredis==2.0.1

# ==========================================
# DATABASE & STORAGE
# ==========================================
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
pymilvus==2.3.3

# ==========================================
# AI/ML & AUDIO PROCESSING
# ==========================================
openai==1.3.0
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0
librosa==0.10.1
soundfile==0.12.1
webrtcvad==2.0.10
numpy==1.24.3
scipy==1.11.3
scikit-learn==1.3.2

# ==========================================
# TELEPHONY & INTEGRATIONS
# ==========================================
twilio==8.10.0
elevenlabs==0.2.26

# ==========================================
# SECURITY & MONITORING
# ==========================================
cryptography==41.0.7
httpx==0.25.2
sentry-sdk==1.38.0
prometheus-client==0.19.0

# ==========================================
# UTILITIES
# ==========================================
python-dotenv==1.0.0
structlog==23.2.0
tenacity==8.2.3
phonenumbers==8.13.26
regex==2023.10.3
pendulum==2.1.2

# ==========================================
# TESTING (Development only)
# ==========================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0
faker==20.1.0
